+++ pomcorn/constants.py
@@ -1,0 +1,8 @@
+from enum import StrEnum
+
+
+class AutomationPlatform(StrEnum):
+    """Enumeration of supported automation platforms."""
+
+    WEB = "web"
+    ANDROID = "android"